+++ src/store.js
@@ -63,7 +63,7 @@ Project*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2022-January-02';
+modules.store = '2022-February-28';
 
 // XML_Serializer ///////////////////////////////////////////////////////
 /*
@@ -1189,7 +1189,8 @@ SnapSerializer.prototype.loadComment = function (model) {
 SnapSerializer.prototype.loadBlock = function (model, isReporter, object) {
     // private
     var block, info, inputs, isGlobal, receiver, migration,
-        migrationOffset = 0;
+        migrationOffset = 0,
+        migratoToVariadic = false; // +++
 
     if (model.tag === 'block') {
         if (Object.prototype.hasOwnProperty.call(
@@ -1206,6 +1207,7 @@ SnapSerializer.prototype.loadBlock = function (model, isReporter, object) {
             ];
             if (migration) {
                 migrationOffset = migration.offset || 0;
+                migratoToVariadic = migration.variadic; // +++
             }
         }
     } else if (model.tag === 'custom-block') {
@@ -1265,7 +1267,14 @@ SnapSerializer.prototype.loadBlock = function (model, isReporter, object) {
         } else if (child.tag === 'receiver') {
             nop(); // ignore
         } else {
-            this.loadInput(child, inputs[i + migrationOffset], block, object);
+            if (migratoToVariadic) { // +++
+                // assume all formerly single inputs are now part of the first
+                // one which is variadic and already expanded to hold them
+                // example: migrate old infix addition to new variadic infix sum
+                this.loadInput(child, inputs[0].inputs()[i], inputs[0], object); // +++
+            } else {
+                this.loadInput(child, inputs[i + migrationOffset], block, object); // +++
+            }
         }
     });
     block.cachedInputs = null;